fix: 기존 세션 클래스 위치 변경 및 유틸리티 메서드 추가
